% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{add_rejections}
\alias{add_rejections}
\title{Mark significant proteins}
\usage{
add_rejections(diff, alpha = 0.05, lfc = 1)
}
\arguments{
\item{diff}{SummarizedExperiment,
Proteomics dataset on which differential enrichment analysis
has been performed (output from \code{\link{test_diff}()}).}

\item{alpha}{Numeric(1),
Sets the threshold for the adjusted P value.}

\item{lfc}{Numeric(1),
Sets the threshold for the log2 fold change.}
}
\value{
A SummarizedExperiment object
annotated with logical columns indicating significant proteins.
}
\description{
\code{add_rejections} marks significant proteins based on defined cutoffs.
}
