% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragPipeAnalystR-package.R
\docType{package}
\name{FragPipeAnalystR-package}
\alias{FragPipeAnalystR}
\alias{FragPipeAnalystR-package}
\title{FragPipeAnalystR: FragPipe downstream analysis in R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
