% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragPipeR-package.R
\docType{package}
\name{FragPipeR-package}
\alias{FragPipeR}
\alias{FragPipeR-package}
\title{FragPipeR: FragPipe downstream analysis in R}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
