% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{get_cluster_heatmap}
\alias{get_cluster_heatmap}
\title{Plot heatmap using differentially expressed features}
\usage{
get_cluster_heatmap(
  dep,
  type = c("contrast", "centered"),
  kmeans = FALSE,
  k = 6,
  col_limit = 6,
  indicate = NULL,
  alpha = 0.01,
  lfc = 1,
  clustering_distance = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski", "pearson", "spearman", "kendall", "gower"),
  row_font_size = 6,
  col_font_size = 10,
  plot = TRUE,
  ...
)
}
\description{
Plot heatmap using differentially expressed features
}
