% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Plot correlation heatmap}
\usage{
plot_correlation_heatmap(
  dep,
  significant = FALSE,
  lower = -1,
  upper = 1,
  pal = "PRGn",
  pal_rev = FALSE,
  indicate = NULL,
  font_size = 10,
  exp = NULL,
  use = "complete.obs",
  ...
)
}
\description{
Plot correlation heatmap
}
