% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccrcc}
\alias{ccrcc}
\title{ccrcc}
\format{
## `ccrcc`
A SummarizedExperiment object from FragPipe-Analyst:
\describe{
}
}
\source{
<https://github.com/Nesvilab/FragPipe-Analyst>
}
\usage{
ccrcc
}
\description{
3-plex CPTAC ccRCC TMT data
}
\keyword{datasets}
